Added ImmutableRoverStatusPacket (that was a lot of typing

+++ src/main/java/me/retrodaredevil/solarthing/solar/renogy/rover/special/UpperLower16Bit.java
@@ -9,7 +9,7 @@ public interface UpperLower16Bit {
 	default int getCombined(){
 		return getCombined(getUpper(), getLower());
 	}
-	default int getCombinedShort(){
+	default short getCombinedShort(){
 		return (short) getCombined();
 	}
 	static int getCombined(int upper, int lower){
